Refactor ConfigurationExtensions for improved value resolution
Significantly refactored `ConfigurationExtensions.cs` by changing the namespace to `ApplicationBuilderHelpers.Extensions`. Introduced a new method `TryGetRefValue` to handle reference chains and improve error handling. Updated existing methods to utilize `TryGetRefValue`, reducing code duplication. Enhanced comments for better documentation of functionality.
